Refractory and have to test first tippool calculation

+++ src/main/java/com/tipout/Tipout/models/Employees/BOH.java
@@ -3,12 +3,13 @@ package com.tipout.Tipout.models.Employees;
 import com.tipout.Tipout.models.Employer;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 public class BOH extends NonMoneyHandler {
     @ManyToOne
     private Employer employer;
-    Integer percentOfTipOut = 2;
+    BigDecimal percentOfTipOut = BigDecimal.valueOf(2);
 
     public BOH() {}
