+++ src/main/java/com/tipout/Tipout/models/Employees/Bartender.java
@@ -1,12 +1,13 @@
 package com.tipout.Tipout.models.Employees;
 
 import javax.persistence.Entity;
+import java.math.BigDecimal;
 
 @Entity
 public class Bartender extends MoneyHandler {
 
 
-    private Integer percentOfTipOut = 10;
+    private BigDecimal percentOfTipOut = BigDecimal.valueOf(10);
 
     public Bartender() {}
