+++ src/main/java/com/tipout/Tipout/models/Employees/Server.java
@@ -1,12 +1,13 @@
 package com.tipout.Tipout.models.Employees;
 
 import javax.persistence.Entity;
+import java.math.BigDecimal;
 
 @Entity
 public class Server extends MoneyHandler {
 
 
-    private Integer percentOfTipOut = 85;
+    private BigDecimal percentOfTipOut = BigDecimal.valueOf(85);
     public Server() {}
 
     public Server(String firstName, String lastName) {
